samples/function-storage/src/requirements.txt
azure-functions
azure-storage-blob>=12.0.0
azure-storage-queue>=12.0.0
azure-data-tables>=12.0.0
azure-identity>=1.12.0
